Various fixes and improvements
Notably return the identity certification which is understood by the
Service Worker. If the identity does not exist or is not certified
executes the previous behaviour. This should help with passing the
e2e tests from dfx.

Also moved the authorization check to use the guard attribute of the
proc_macro (so its rejected), and allowing to return uncertified
assets (not used right now).