fix: trust full asset SHA-256 hashes provided by the caller
This change removes abligatory SHA-256 recomputation when a full asset
is assembled out of chunks.  The hash is only recomputed if it wasn't
specified by the caller.

Computing the full hash causes the canister to run out of cycles when
large assets are uploaded (~50MiB. Yes, there are people who do this).

Fixes #5.